backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.4",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
]
